Move size selectors to separate file
