+++ packages/grid/_modules_/grid/hooks/utils/sizesSelector.ts
@@ -1,0 +1,5 @@
+import { GridState } from '../features/core/gridState';
+
+export const gridContainerSizesSelector = (state: GridState) => state.containerSizes;
+export const gridViewportSizesSelector = (state: GridState) => state.viewportSizes;
+export const gridScrollBarSizeSelector = (state: GridState) => state.scrollBar;
